Dedup a bunch of tests
