+++ crates/vm/src/arch/testing/cpu.rs
@@ -149,6 +149,10 @@ where
     fn execution_final_state(&self) -> ExecutionState<F> {
         self.execution.records.last().unwrap().final_state
     }
+
+    fn streams_mut(&mut self) -> &mut Streams<F> {
+        &mut self.streams
+    }
 }
 
 impl<F: PrimeField32> VmChipTestBuilder<F> {
