+++ crates/vm/src/arch/testing/mod.rs
@@ -19,7 +19,7 @@ use openvm_stark_backend::{interaction::BusIndex, p3_air::BaseAir};
 use p3_field::Field;
 pub use utils::*;
 
-use crate::arch::{Arena, ExecutionState, MatrixRecordArena, PreflightExecutor};
+use crate::arch::{Arena, ExecutionState, MatrixRecordArena, PreflightExecutor, Streams};
 
 pub const EXECUTION_BUS: BusIndex = 0;
 pub const MEMORY_BUS: BusIndex = 1;
@@ -88,4 +88,5 @@ pub trait TestBuilder<F> {
     );
 
     fn execution_final_state(&self) -> ExecutionState<F>;
+    fn streams_mut(&mut self) -> &mut Streams<F>;
 }
