+++ extensions/native/circuit/src/field_arithmetic/cuda.rs
@@ -55,119 +55,3 @@ impl Chip<DenseRecordArena, GpuBackend> for FieldArithmeticChipGpu {
         AirProvingContext::simple_no_pis(trace)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use openvm_circuit::arch::{
-        testing::{memory::gen_pointer, GpuChipTestBuilder, GpuTestChipHarness, TestBuilder},
-        EmptyAdapterCoreLayout,
-    };
-    use openvm_instructions::{instruction::Instruction, LocalOpcode};
-    use openvm_native_compiler::{conversion::AS, FieldArithmeticOpcode};
-    use openvm_stark_backend::p3_field::{Field, FieldAlgebra};
-    use openvm_stark_sdk::utils::create_seeded_rng;
-    use rand::Rng;
-    use test_case::test_case;
-
-    use super::*;
-    use crate::{
-        adapters::{AluNativeAdapterAir, AluNativeAdapterExecutor, AluNativeAdapterFiller},
-        field_arithmetic::{
-            FieldArithmeticAir, FieldArithmeticChip, FieldArithmeticCoreAir,
-            FieldArithmeticCoreFiller, FieldArithmeticExecutor,
-        },
-        write_native_array,
-    };
-
-    const MAX_INS_CAPACITY: usize = 128;
-
-    fn create_test_harness(
-        tester: &GpuChipTestBuilder,
-    ) -> GpuTestChipHarness<
-        F,
-        FieldArithmeticExecutor,
-        FieldArithmeticAir,
-        FieldArithmeticChipGpu,
-        FieldArithmeticChip<F>,
-    > {
-        let adapter_air =
-            AluNativeAdapterAir::new(tester.execution_bridge(), tester.memory_bridge());
-        let core_air = FieldArithmeticCoreAir::new();
-        let air = FieldArithmeticAir::new(adapter_air, core_air);
-
-        let adapter_step = AluNativeAdapterExecutor::new();
-        let executor = FieldArithmeticExecutor::new(adapter_step);
-
-        let core_filler = FieldArithmeticCoreFiller::new(AluNativeAdapterFiller);
-
-        let cpu_chip = FieldArithmeticChip::new(core_filler, tester.dummy_memory_helper());
-        let gpu_chip =
-            FieldArithmeticChipGpu::new(tester.range_checker(), tester.timestamp_max_bits());
-
-        GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-    }
-
-    #[test_case(FieldArithmeticOpcode::ADD, 100)]
-    #[test_case(FieldArithmeticOpcode::SUB, 100)]
-    #[test_case(FieldArithmeticOpcode::MUL, 100)]
-    #[test_case(FieldArithmeticOpcode::DIV, 100)]
-    fn rand_field_arithmetic_tracegen_test(opcode: FieldArithmeticOpcode, num_ops: usize) {
-        let mut tester = GpuChipTestBuilder::default();
-        let mut rng = create_seeded_rng();
-
-        let mut harness = create_test_harness(&tester);
-
-        for _ in 0..num_ops {
-            let b_val = rng.gen::<F>();
-            let c_val = if opcode == FieldArithmeticOpcode::DIV {
-                loop {
-                    let x = rng.gen::<F>();
-                    if !x.is_zero() {
-                        break x;
-                    }
-                }
-            } else {
-                rng.gen::<F>()
-            };
-
-            let (_, b_ptr) = write_native_array(&mut tester, &mut rng, Some([b_val]));
-            let (_, c_ptr) = write_native_array(&mut tester, &mut rng, Some([c_val]));
-
-            let a = gen_pointer(&mut rng, 1);
-            tester.execute(
-                &mut harness.executor,
-                &mut harness.dense_arena,
-                &Instruction::new(
-                    opcode.global_opcode(),
-                    F::from_canonical_usize(a),
-                    F::from_canonical_usize(b_ptr),
-                    F::from_canonical_usize(c_ptr),
-                    F::from_canonical_usize(AS::Native as usize),
-                    F::from_canonical_usize(AS::Native as usize),
-                    F::from_canonical_usize(AS::Native as usize),
-                    F::ZERO,
-                ),
-            );
-        }
-
-        type Record<'a> = (
-            &'a mut AluNativeAdapterRecord<F>,
-            &'a mut FieldArithmeticRecord<F>,
-        );
-
-        harness
-            .dense_arena
-            .get_record_seeker::<Record<'_>, _>()
-            .transfer_to_matrix_arena(
-                &mut harness.matrix_arena,
-                EmptyAdapterCoreLayout::<F, AluNativeAdapterExecutor>::new(),
-            );
-
-        tester
-            .build()
-            .load_gpu_harness(harness)
-            .finalize()
-            .simple_test()
-            .unwrap();
-    }
-}
