+++ extensions/native/circuit/src/field_extension/cuda.rs
@@ -56,124 +56,3 @@ impl Chip<DenseRecordArena, GpuBackend> for FieldExtensionChipGpu {
         AirProvingContext::simple_no_pis(trace)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use openvm_circuit::arch::{
-        testing::{memory::gen_pointer, GpuChipTestBuilder, GpuTestChipHarness, TestBuilder},
-        EmptyAdapterCoreLayout,
-    };
-    use openvm_instructions::{instruction::Instruction, LocalOpcode};
-    use openvm_native_compiler::{conversion::AS, FieldExtensionOpcode};
-    use openvm_stark_sdk::utils::create_seeded_rng;
-    use rand::rngs::StdRng;
-    use test_case::test_case;
-
-    use super::*;
-    use crate::{
-        adapters::{
-            NativeVectorizedAdapterAir, NativeVectorizedAdapterExecutor,
-            NativeVectorizedAdapterFiller,
-        },
-        field_extension::{
-            FieldExtensionAir, FieldExtensionChip, FieldExtensionCoreAir, FieldExtensionCoreFiller,
-            FieldExtensionExecutor,
-        },
-        write_native_array,
-    };
-    const MAX_INS_CAPACITY: usize = 128;
-
-    fn create_test_harness(
-        tester: &GpuChipTestBuilder,
-    ) -> GpuTestChipHarness<
-        F,
-        FieldExtensionExecutor,
-        FieldExtensionAir,
-        FieldExtensionChipGpu,
-        FieldExtensionChip<F>,
-    > {
-        let adapter_air =
-            NativeVectorizedAdapterAir::new(tester.execution_bridge(), tester.memory_bridge());
-        let core_air = FieldExtensionCoreAir::new();
-        let air = FieldExtensionAir::new(adapter_air, core_air);
-
-        let adapter_step = NativeVectorizedAdapterExecutor::new();
-        let executor = FieldExtensionExecutor::new(adapter_step);
-
-        let core_filler = FieldExtensionCoreFiller::new(NativeVectorizedAdapterFiller);
-
-        let cpu_chip = FieldExtensionChip::new(core_filler, tester.dummy_memory_helper());
-        let gpu_chip =
-            FieldExtensionChipGpu::new(tester.range_checker(), tester.timestamp_max_bits());
-
-        GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-    }
-
-    fn set_and_execute(
-        tester: &mut GpuChipTestBuilder,
-        harness: &mut GpuTestChipHarness<
-            F,
-            FieldExtensionExecutor,
-            FieldExtensionAir,
-            FieldExtensionChipGpu,
-            FieldExtensionChip<F>,
-        >,
-        rng: &mut StdRng,
-        opcode: FieldExtensionOpcode,
-    ) {
-        let (_, y_ptr) = write_native_array::<F, EXT_DEG>(tester, rng, None);
-        let (_, z_ptr) = write_native_array::<F, EXT_DEG>(tester, rng, None);
-
-        let x_ptr = gen_pointer(rng, EXT_DEG);
-
-        tester.execute(
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &Instruction::from_usize(
-                opcode.global_opcode(),
-                [
-                    x_ptr,
-                    y_ptr,
-                    z_ptr,
-                    AS::Native as usize,
-                    AS::Native as usize,
-                ],
-            ),
-        );
-    }
-
-    #[test_case(FieldExtensionOpcode::FE4ADD, 100)]
-    #[test_case(FieldExtensionOpcode::FE4SUB, 100)]
-    #[test_case(FieldExtensionOpcode::BBE4MUL, 100)]
-    #[test_case(FieldExtensionOpcode::BBE4DIV, 100)]
-    fn rand_field_extension_tracegen_test(opcode: FieldExtensionOpcode, num_ops: usize) {
-        let mut rng = create_seeded_rng();
-        let mut tester = GpuChipTestBuilder::default();
-
-        let mut harness = create_test_harness(&tester);
-
-        for _ in 0..num_ops {
-            set_and_execute(&mut tester, &mut harness, &mut rng, opcode);
-        }
-
-        type Record<'a> = (
-            &'a mut NativeVectorizedAdapterRecord<F, EXT_DEG>,
-            &'a mut FieldExtensionRecord<F>,
-        );
-
-        harness
-            .dense_arena
-            .get_record_seeker::<Record<'_>, _>()
-            .transfer_to_matrix_arena(
-                &mut harness.matrix_arena,
-                EmptyAdapterCoreLayout::<F, NativeVectorizedAdapterExecutor<EXT_DEG>>::new(),
-            );
-
-        tester
-            .build()
-            .load_gpu_harness(harness)
-            .finalize()
-            .simple_test()
-            .unwrap();
-    }
-}
