+++ extensions/native/circuit/src/fri/tests.rs
@@ -1,8 +1,9 @@
 use std::borrow::BorrowMut;
 
 use itertools::Itertools;
-use openvm_circuit::arch::testing::{
-    memory::gen_pointer, TestBuilder, TestChipHarness, VmChipTestBuilder,
+use openvm_circuit::arch::{
+    testing::{memory::gen_pointer, TestBuilder, TestChipHarness, VmChipTestBuilder},
+    Arena, PreflightExecutor,
 };
 use openvm_instructions::{instruction::Instruction, LocalOpcode};
 use openvm_native_compiler::{conversion::AS, FriOpcode::FRI_REDUCED_OPENING};
@@ -17,15 +18,22 @@ use openvm_stark_backend::{
 };
 use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
 use rand::{rngs::StdRng, Rng};
+use test_case::test_case;
 
 use super::{
     super::field_extension::FieldExtension, elem_to_ext, FriReducedOpeningAir,
     FriReducedOpeningChip, FriReducedOpeningExecutor, EXT_DEG,
 };
+#[cfg(feature = "cuda")]
+use crate::fri::cuda::FriReducedOpeningChipGpu;
 use crate::{
     fri::{FriReducedOpeningFiller, WorkloadCols, OVERALL_WIDTH, WL_WIDTH},
     write_native_array,
 };
+#[cfg(feature = "cuda")]
+use openvm_circuit::arch::testing::GpuChipTestBuilder;
+#[cfg(feature = "cuda")]
+use openvm_circuit::arch::testing::GpuTestChipHarness;
 
 const MAX_INS_CAPACITY: usize = 1024;
 type F = BabyBear;
@@ -38,6 +46,27 @@ fn create_test_chip(tester: &VmChipTestBuilder<F>) -> Harness {
     let chip = FriReducedOpeningChip::new(FriReducedOpeningFiller, tester.memory_helper());
 
     Harness::with_capacity(step, air, chip, MAX_INS_CAPACITY)
+}
+
+#[cfg(feature = "cuda")]
+fn create_test_harness(
+    tester: &GpuChipTestBuilder,
+) -> GpuTestChipHarness<
+    F,
+    FriReducedOpeningExecutor,
+    FriReducedOpeningAir,
+    FriReducedOpeningChipGpu,
+    FriReducedOpeningChip<F>,
+> {
+    let air = FriReducedOpeningAir::new(tester.execution_bridge(), tester.memory_bridge());
+    let executor = FriReducedOpeningExecutor;
+
+    let cpu_chip =
+        FriReducedOpeningChip::new(FriReducedOpeningFiller, tester.dummy_memory_helper());
+    let gpu_chip =
+        FriReducedOpeningChipGpu::new(tester.range_checker(), tester.timestamp_max_bits());
+
+    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
 }
 
 fn compute_fri_mat_opening<F: Field>(
@@ -57,7 +86,15 @@ fn compute_fri_mat_opening<F: Field>(
     result
 }
 
-fn set_and_execute(tester: &mut VmChipTestBuilder<F>, harness: &mut Harness, rng: &mut StdRng) {
+fn set_and_execute<E, RA>(
+    tester: &mut impl TestBuilder<F>,
+    executor: &mut E,
+    arena: &mut RA,
+    rng: &mut StdRng,
+) where
+    E: PreflightExecutor<F, RA>,
+    RA: Arena,
+{
     let len = rng.gen_range(1..=28);
     let a_ptr = gen_pointer(rng, len);
     let b_ptr = gen_pointer(rng, len);
@@ -80,7 +117,7 @@ fn set_and_execute(tester: &mut VmChipTestBuilder<F>, harness: &mut Harness, rng
         vec_a.push(a);
         vec_b.push(b);
         if !is_init {
-            tester.streams.hint_space[0].push(a);
+            tester.streams_mut().hint_space[0].push(a);
         } else {
             tester.write(AS::Native as usize, a_ptr + i, [a]);
         }
@@ -88,8 +125,8 @@ fn set_and_execute(tester: &mut VmChipTestBuilder<F>, harness: &mut Harness, rng
     }
 
     tester.execute(
-        &mut harness.executor,
-        &mut harness.arena,
+        executor,
+        arena,
         &Instruction::from_usize(
             FRI_REDUCED_OPENING.global_opcode(),
             [
@@ -128,11 +165,40 @@ fn fri_mat_opening_air_test() {
 
     let num_ops = 28; // non-power-of-2 to also test padding
     for _ in 0..num_ops {
-        set_and_execute(&mut tester, &mut harness, &mut rng);
+        set_and_execute(
+            &mut tester,
+            &mut harness.executor,
+            &mut harness.arena,
+            &mut rng,
+        );
     }
 
     let tester = tester.build().load(harness).finalize();
     tester.simple_test().expect("Verification failed");
+}
+
+#[cfg(feature = "cuda")]
+#[test_case(28)]
+fn test_fri_tracegen(num_ops: usize) {
+    let mut rng = create_seeded_rng();
+    let mut tester = GpuChipTestBuilder::default();
+    let mut harness = create_test_harness(&tester);
+
+    for _ in 0..num_ops {
+        set_and_execute(
+            &mut tester,
+            &mut harness.executor,
+            &mut harness.dense_arena,
+            &mut rng,
+        );
+    }
+
+    tester
+        .build()
+        .load_gpu_harness(harness)
+        .finalize()
+        .simple_test()
+        .unwrap();
 }
 
 //////////////////////////////////////////////////////////////////////////////////////
@@ -148,7 +214,12 @@ fn run_negative_fri_mat_opening_test() {
     let mut tester = VmChipTestBuilder::default_native();
     let mut harness = create_test_chip(&tester);
 
-    set_and_execute(&mut tester, &mut harness, &mut rng);
+    set_and_execute(
+        &mut tester,
+        &mut harness.executor,
+        &mut harness.arena,
+        &mut rng,
+    );
 
     let modify_trace = |trace: &mut DenseMatrix<F>| {
         let mut values = trace.row_slice(0).to_vec();
