+++ extensions/native/circuit/src/jal_rangecheck/cuda.rs
@@ -50,108 +50,3 @@ impl Chip<DenseRecordArena, GpuBackend> for JalRangeCheckGpu {
         AirProvingContext::simple_no_pis(trace)
     }
 }
-
-#[cfg(test)]
-mod test {
-    use openvm_circuit::{
-        arch::{
-            testing::{memory::gen_pointer, GpuChipTestBuilder, GpuTestChipHarness, TestBuilder},
-            EmptyMultiRowLayout,
-        },
-        utils::cuda_utils::{default_var_range_checker_bus, dummy_range_checker},
-    };
-    use openvm_instructions::{instruction::Instruction, program::PC_BITS, LocalOpcode, VmOpcode};
-    use openvm_native_compiler::{conversion::AS, NativeJalOpcode, NativeRangeCheckOpcode};
-    use openvm_stark_backend::p3_field::FieldAlgebra;
-    use openvm_stark_sdk::utils::create_seeded_rng;
-    use rand::Rng;
-    use test_case::test_case;
-
-    use super::*;
-    use crate::jal_rangecheck::{
-        JalRangeCheckAir, JalRangeCheckExecutor, JalRangeCheckFiller, NativeJalRangeCheckChip,
-    };
-
-    const MAX_INS_CAPACITY: usize = 128;
-
-    fn create_test_harness(
-        tester: &GpuChipTestBuilder,
-    ) -> GpuTestChipHarness<
-        F,
-        JalRangeCheckExecutor,
-        JalRangeCheckAir,
-        JalRangeCheckGpu,
-        NativeJalRangeCheckChip<F>,
-    > {
-        let range_bus = default_var_range_checker_bus();
-        let air =
-            JalRangeCheckAir::new(tester.execution_bridge(), tester.memory_bridge(), range_bus);
-        let executor = JalRangeCheckExecutor::new();
-        let filler = JalRangeCheckFiller::new(dummy_range_checker(range_bus));
-        let cpu_chip = NativeJalRangeCheckChip::<F>::new(filler, tester.dummy_memory_helper());
-        let gpu_chip = JalRangeCheckGpu::new(tester.range_checker(), tester.timestamp_max_bits());
-
-        GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-    }
-
-    #[test_case(NativeJalOpcode::JAL.global_opcode(), 100)]
-    #[test_case(NativeRangeCheckOpcode::RANGE_CHECK.global_opcode(), 100)]
-    fn test_jal_rangecheck_tracegen(opcode: VmOpcode, num_ops: usize) {
-        let mut rng = create_seeded_rng();
-        let mut tester = GpuChipTestBuilder::default();
-        let mut harness = create_test_harness(&tester);
-
-        for _ in 0..num_ops {
-            if opcode == NativeJalOpcode::JAL.global_opcode() {
-                let initial_pc = rng.gen_range(0..(1 << PC_BITS));
-                let a = gen_pointer(&mut rng, 1);
-                let b = rng.gen_range(0..(1 << 16));
-
-                tester.execute_with_pc(
-                    &mut harness.executor,
-                    &mut harness.dense_arena,
-                    &Instruction::from_usize(
-                        opcode,
-                        [a, b as usize, 0, AS::Native as usize, 0, 0, 0],
-                    ),
-                    initial_pc,
-                );
-            } else {
-                // RANGE_CHECK
-                let a_val = rng.gen_range(0..(1 << 30));
-                let a = gen_pointer(&mut rng, 1);
-                tester.write::<1>(AS::Native as usize, a, [F::from_canonical_u32(a_val)]);
-
-                let x = a_val & 0xffff;
-                let y = a_val >> 16;
-                let min_b = 32 - x.leading_zeros();
-                let min_c = 32 - y.leading_zeros();
-                let b = rng.gen_range(min_b..=16);
-                let c = rng.gen_range(min_c..=14);
-
-                tester.execute(
-                    &mut harness.executor,
-                    &mut harness.dense_arena,
-                    &Instruction::from_usize(
-                        opcode,
-                        [a, b as usize, c as usize, AS::Native as usize, 0, 0, 0],
-                    ),
-                );
-            }
-        }
-
-        type Record<'a> = &'a mut JalRangeCheckRecord<F>;
-
-        harness
-            .dense_arena
-            .get_record_seeker::<Record<'_>, EmptyMultiRowLayout>()
-            .transfer_to_matrix_arena(&mut harness.matrix_arena);
-
-        tester
-            .build()
-            .load_gpu_harness(harness)
-            .finalize()
-            .simple_test()
-            .unwrap();
-    }
-}
